Build out costs. Start dealing with s wrap around. More tests

+++ src/cost.hpp
@@ -5,7 +5,8 @@
 
 namespace cost {
 
-float
-calculateCost(const Vehicle &ego, const prediction::Predictions &predictions, const std::vector<Vehicle> &trajectory);
+double calculateCost(const Vehicle &ego,
+                     const prediction::Predictions &predictions,
+                     const std::vector<Vehicle> &trajectory);
 
 }
